Removed static_cast as approx uses double anyway. Fixing a typo in comments.

+++ src/mlpack/tests/main_tests/emst_test.cpp
@@ -104,10 +104,10 @@ TEST_CASE_METHOD(EMSTTestFixture, "EMSTFirstTwoOutputRowsIntegerTest",
   for (size_t i = 0; i < params.Get<arma::mat>("output").n_cols; ++i)
   {
     REQUIRE(params.Get<arma::mat>("output")(0, i) ==
-        Approx(static_cast<int>(std::round(params.Get<arma::mat>("output")(0, i)))).
+        Approx(std::round(params.Get<arma::mat>("output")(0, i))).
         epsilon(1e-7));
     REQUIRE(params.Get<arma::mat>("output")(1, i) ==
-        Approx(static_cast<int>(std::round(params.Get<arma::mat>("output")(1, i)))).
+        Approx(std::round(params.Get<arma::mat>("output")(1, i))).
         epsilon(1e-7));
   }
 }
